+++ src/clients/client/ClientOptions.ts
@@ -1,12 +1,12 @@
 import { Storage } from '../..'
-import { BeaconEvent, BeaconEventHandlerFunction } from '../../events'
+import { BeaconEvent, BeaconEventHandlerFunction, BeaconEventType } from '../../events'
 
 export interface ClientOptions {
   name: string
   storage: Storage
   eventHandlers?: {
     [key in BeaconEvent]?: {
-      handler: BeaconEventHandlerFunction
+      handler: BeaconEventHandlerFunction<BeaconEventType[key]>
     }
   }
 }