+++ src/types/PermissionEntity.ts
@@ -1,0 +1,7 @@
+import { Network, PermissionScope } from '..'
+
+export interface PermissionEntity {
+  address: string
+  network: Network
+  scopes: PermissionScope[]
+}