+++ src/types/PermissionInfo.ts
@@ -1,0 +1,11 @@
+import { AppMetadata } from '..'
+import { PermissionEntity } from './PermissionEntity'
+
+export interface PermissionInfo extends PermissionEntity {
+  accountIdentifier: string
+  beaconId: string
+  appMetadata: AppMetadata
+  website: string
+  publicKey: string
+  connectedAt: number
+}