+++ src/types/storage/StorageKey.ts
@@ -2,5 +2,8 @@ export enum StorageKey {
   TRANSPORT_P2P_PEERS = 'beacon:communication-peers',
   ACCOUNTS = 'beacon:accounts',
   ACTIVE_ACCOUNT = 'beacon:active-account',
-  BEACON_SDK_SECRET_SEED = 'beacon:sdk-secret-seed'
+  BEACON_SDK_SECRET_SEED = 'beacon:sdk-secret-seed',
+  APP_METADATA_LIST = 'beacon:app-metadata-list',
+  PERMISSION_LIST = 'beacon:permissions',
+  BEACON_SDK_VERSION = 'beacon:sdk_version'
 }