+++ src/types/storage/StorageKeyReturnDefaults.ts
@@ -6,5 +6,8 @@ export const defaultValues: StorageKeyReturnDefaults = {
   [StorageKey.TRANSPORT_P2P_PEERS]: [],
   [StorageKey.ACCOUNTS]: [],
   [StorageKey.ACTIVE_ACCOUNT]: undefined,
-  [StorageKey.BEACON_SDK_SECRET_SEED]: undefined
+  [StorageKey.BEACON_SDK_SECRET_SEED]: undefined,
+  [StorageKey.APP_METADATA_LIST]: [],
+  [StorageKey.PERMISSION_LIST]: [],
+  [StorageKey.BEACON_SDK_VERSION]: undefined
 }